Fix placeholder Atom feed (#121)
The Jekyll-feed plugin wasn't actually working because there was a static, placeholder `feed.xml` file. Removing the file allows the plugin to do its work. That said, it looks like there are no items in the feed because none of the pages in this site are *posts*, so... debatable value here.