Add banner indicating this is from Trump 1
Now that we are in a second Trump term (and in the season of publishing reports like this first-six-months one), some folks might land on these pages thinking they are about the current administration instead of the first one in 2017. This adds a small banner to the top of every page letting people know that these reports are about 2017 and linking to EDGI's blog for more current info.

Fixes #120.