Add banner indicating this is from Trump 1 (#124)
Now that we are in a second Trump term (and in the season of publishing reports like this first-six-months one), people might land on these pages thinking they are about the current administration instead of the first one in 2017. This adds a small banner to the top of every page letting people know that these reports are about 2017 and linking to EDGI's blog for more current info.

The banner just sits at the top of the page and is NOT sticky (does not stay at the top of the window when you scroll).

The styles are directly in the HTML here because it turns out this site used two totally separate stylesheets for the home page vs. other pages, and there’s no built-in way to share common styles across them (!).